apps/sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentguard-protocol"
version = "0.1.0"
description = "Governance layer SDK for autonomous AI agent payments on Circle Nanopayments and Arc"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "AgentGuard Protocol", email = "hi@agentguard.dev"},
]
readme = "README.md"
keywords = [
  "ai-agents",
  "agent-payments",
  "circle",
  "usdc",
  "arc",
  "nanopayments",
  "erc-8004",
  "policy-engine",
  "governance",
  "a2a",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.5",
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/agentguard/agentguard"
Documentation = "https://github.com/agentguard/agentguard#readme"
Repository = "https://github.com/agentguard/agentguard"
Issues = "https://github.com/agentguard/agentguard/issues"
Changelog = "https://github.com/agentguard/agentguard/blob/main/CHANGELOG.md"

[project.optional-dependencies]
# Enables real Circle Gateway + x402 payments via circlekit.
# Install with: pip install "agentguard-protocol[circle]"
circle = [
  "circle-titanoboa-sdk>=0.1.0",
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-httpx>=0.30",
  "ruff>=0.4",
  "mypy>=1.8",
]

[tool.hatch.build.targets.wheel]
packages = ["agentguard"]

[tool.hatch.build.targets.sdist]
include = [
  "/agentguard",
  "/tests",
  "/README.md",
  "/CHANGELOG.md",
  "/pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = false
